+++ src/main/java/roomescape/web/controller/MemberController.java
@@ -5,7 +5,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.dto.member.MemberResponse;
-import roomescape.service.MemberService;
+import roomescape.service.member.MemberService;
 
 @RestController
 class MemberController {
