+++ src/main/java/roomescape/web/interceptor/AdminAuthorizationInterceptor.java
@@ -8,7 +8,7 @@ import roomescape.domain.member.Role;
 import roomescape.dto.login.LoginMember;
 import roomescape.dto.token.TokenDto;
 import roomescape.infrastructure.auth.AuthorizationExtractor;
-import roomescape.service.AuthService;
+import roomescape.service.auth.AuthService;
 
 @Component
 public class AdminAuthorizationInterceptor implements HandlerInterceptor {
